+++ CHANGELOG.md
@@ -24,6 +24,7 @@ For details about compatibility between different releases, see the **Commitment
 - Uplink frame counter reset handling.
 - Uplink retransmission handling in Network Server.
 - DevAddr generation for NetID Type 3 and 4, according to errata.
+- HTTP header propagation (such as Request ID) to gRPC services.
 
 ### Security